requirements.txt
python-decouple==3.8
pyodbc==5.2.0
sqlalchemy==2.0.32
pydantic==2.11.8